requirements.txt
fastapi
uvicorn
boto3
pinecone
marker-pdf
pdfplumber
tiktoken
python-dotenv
requests
